requirements.txt
sympy
mpmath
numpy
matplotlib
pybind11
scipy
pandas
google-generativeai
openai
python-dotenv
